requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
timm
fastapi
uvicorn
python-multipart
albumentations
Pillow
numpy
opencv-python-headless
huggingface_hub
